requirements.txt
streamlit
webbrowser
Pillow